+++ commercetools/commercetools-sdk-java-importapi/src/main/java-generated/com/commercetools/importapi/models/customers/CustomerImport.java
@@ -64,7 +64,7 @@ public interface CustomerImport extends ImportResource {
     public String getEmail();
 
     /**
-     *  <p>Maps to <code>Customer.password</code>. Required when <code>authenticationMode</code> is set to <code>Password</code>.</p>
+     *  <p>Maps to <code>Customer.password</code>. Required when <code>authenticationMode</code> is set to <code>Password</code>. This field is only used when creating new Customers, and is ignored when updating Customers. However, due to the Import API's upsert behavior, a non-empty value must still be provided when updating Customers.</p>
      * @return password
      */
 
@@ -256,7 +256,7 @@ public interface CustomerImport extends ImportResource {
     public void setEmail(final String email);
 
     /**
-     *  <p>Maps to <code>Customer.password</code>. Required when <code>authenticationMode</code> is set to <code>Password</code>.</p>
+     *  <p>Maps to <code>Customer.password</code>. Required when <code>authenticationMode</code> is set to <code>Password</code>. This field is only used when creating new Customers, and is ignored when updating Customers. However, due to the Import API's upsert behavior, a non-empty value must still be provided when updating Customers.</p>
      * @param password value to be set
      */