+++ changes.md
@@ -3,6 +3,7 @@
 <details>
 <summary>Added Type(s)</summary>
 
+- added type `InventoryEntrySetInventoryLimitsAction`
 - added type `RecurringOrderDeletedMessage`
 - added type `RecurringOrderExpiresAtSetMessage`
 - added type `RecurringOrderDeletedMessagePayload`
@@ -32,6 +33,10 @@
 <details>
 <summary>Added Property(s)</summary>
 
+- added property `minCartQuantity` to type `InventoryEntry`
+- added property `maxCartQuantity` to type `InventoryEntry`
+- added property `minCartQuantity` to type `InventoryEntryDraft`
+- added property `maxCartQuantity` to type `InventoryEntryDraft`
 - added property `expiresAt` to type `RecurringOrderDraft`
 </details>